Extract Importer for avoiding code duplication
