+++ scripts/importer.rb
@@ -1,0 +1,89 @@
+# Decomoji Importer
+class Importer
+  DEFAULT_IMPORT_IMG_DIR = File.expand_path(File.dirname(__FILE__)) + "/../decomoji/basic"
+
+  def initialize(import_img_dir: nil)
+    @page = nil
+    @agent = Mechanize.new
+    @import_img_dir = import_img_dir || DEFAULT_IMPORT_IMG_DIR
+  end
+  attr_accessor :page, :agent, :team_name, :token
+
+  def import_decomojis
+    move_to_emoji_page
+    upload_decomojis
+  end
+
+  private
+
+  def login
+    @team_name = ask('Your slack team name(subdomain): ')
+    email      = ask('Login email: ')
+    password   = ask('Login password(hidden): ') { |q| q.echo = false }
+
+    emoji_page_url = "https://#{team_name}.slack.com/customize/emoji"
+
+    page = agent.get(emoji_page_url)
+    page.form.email = email
+    page.form.password = password
+    @page = page.form.submit
+    @token = @page.body[/(?<=api_token":")[^"]+/]
+  end
+
+  def enter_two_factor_authentication_code
+    page.form['2fa_code'] = ask('Your two factor authentication code: ')
+    @page = page.form.submit
+    @token = @page.body[/(?<=api_token":")[^"]+/]
+  end
+
+  def move_to_emoji_page
+    loop do
+      if page && page.form['signin_2fa']
+        enter_two_factor_authentication_code
+      else
+        login
+      end
+
+      break if page.title.include?('絵文字') || page.title.include?('Emoji')
+      puts 'Login failure. Please try again.'
+      puts
+    end
+  end
+
+  def upload_decomojis
+    emojis = list_emojis
+    files = Dir.glob(@import_img_dir + "/*.png")
+    len = files.length
+    files.each.with_index(1) do |path, i|
+      basename = File.basename(path, '.*')
+
+      # skip if already exists
+      if emojis.include?(basename)
+        puts "(#{i}/#{len}) #{basename} already exists, skip"
+        next
+      end
+
+      puts "(#{i}/#{len}) importing #{basename}..."
+
+      params = {
+        name: basename,
+        image: File.new(path),
+        mode: 'data',
+        token: token
+      }
+      agent.post("https://#{team_name}.slack.com/api/emoji.add", params)
+    end
+  end
+
+  def list_emojis
+    emojis = []
+    loop.with_index(1) do |_, n|
+      params = { query: '', page: n, count: 100, token: token }
+      res = JSON.parse(agent.post("https://#{team_name}.slack.com/api/emoji.adminList", params).body)
+      raise res['error'] if res['error']
+      emojis.push(*res['emoji'].map { |e| e['name'] })
+      break if res['paging']['pages'] == n || res['paging']['pages'] == 0
+    end
+    emojis
+  end
+end
